backend/requirements.txt
Django==3.1.8
asgiref>=3.2.10,<4.0
pytz==2019.3
sqlparse==0.3.1

djangorestframework==3.11.2
requests==2.23.0
certifi==2019.11.28
chardet==3.0.4
idna==2.9
urllib3==1.25.8
coreschema==0.0.4
itypes==1.1.0
uritemplate==3.0.1
Jinja2==2.11.3
MarkupSafe==1.1.1
django-cors-headers==3.5.0


psycopg2-binary
gunicorn
python-dotenv
#psycopg2-binary
#psycopg2-binary==2.9.9
